requirements.txt
python-telegram-bot==20.7
feedparser
openai==0.28
aiohttp